+++ src/jobs/CsvRowsJob.php
@@ -57,7 +57,7 @@ class CsvRowsJob extends BaseJob
 
         $optionSiteId = isset($this->export['siteId']) ? $this->export['siteId'] : null;
         // $optionExpireEntries = null; // isset($this->export['expireEntries']) ? count($this->export['expireEntries']) : null;
-        $optionExpireEntries = isset($this->export['expireEntries']) ? count($this->export['expireEntries']) : null;
+        $optionExpireEntries = isset($this->export['expireEntries']) ? $this->export['expireEntries'] : null;
         $requestSite = Craft::$app->sites->getSiteById((int) $optionSiteId) !== null
             ? Craft::$app->sites->getSiteById((int) $optionSiteId)->id
             : Craft::$app->sites->primarySite->id;