+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 5.1.0 - 2024-09-14
+### Added
+- Add ability to specify section handle and field handle to use to exclude entries from export. This is needed to allow us to export those schools that have not placed any FIS orders.
+
 ## 5.0.1 - 2024-08-12
 ### Added
 - getSections is now getEntries
@@ -14,7 +18,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## 4.4.0 - 2024-04-08
 ### Added
-- Allow the entry id to be used in conact handle
+- Allow the entry id to be used in contact handle
 
 ## 4.3.0 - 2024-02-27
 ### Added